fix: resolve require cycle warnings
