+++ src/Root.tsx
@@ -24,8 +24,9 @@ import { useReconnectIndexer } from './hooks/useReconnectIndexer'
 import { logger } from './lib/logger'
 import { ToastProvider } from './lib/toastContext'
 import { uniqueId } from './lib/uniqueId'
+import { initApp, shutdownApp } from './managers/app'
 import { RootTabs } from './stacks/RootTabs'
-import { initApp, shutdownApp, useShowSplash } from './stores/app'
+import { useShowSplash } from './stores/app'
 import { useHasOnboarded } from './stores/settings'
 import { palette } from './styles/colors'
