+++ src/stores/app.ts
@@ -1,36 +1,5 @@
 import { create } from 'zustand'
-import { initializeDB, resetDb } from '../db'
 import { createGetterAndSelector } from '../lib/selectors'
-import { shutdownAllServiceIntervals } from '../lib/serviceInterval'
-import { initBackgroundTasks } from '../managers/backgroundTasks'
-import { initFsEvictionScanner } from '../managers/fsEvictionScanner'
-import { initFsOrphanScanner } from '../managers/fsOrphanScanner'
-import { initLogRotation } from '../managers/logRotation'
-import {
-  initSyncDownEvents,
-  resetSyncDownCursor,
-} from '../managers/syncDownEvents'
-import {
-  initSyncNewPhotos,
-  resetPhotosNewCursor,
-} from '../managers/syncNewPhotos'
-import {
-  initSyncPhotosArchive,
-  resetPhotosArchiveCursor,
-} from '../managers/syncPhotosArchive'
-import { initSyncUpMetadata } from '../managers/syncUpMetadata'
-import { initThumbnailScanner } from '../managers/thumbnailScanner'
-import { initUploadScanner } from '../managers/uploadScanner'
-import { clearAppKeys, migrateKeychainAccessibility } from './appKey'
-import { cancelAllDownloads } from './downloads'
-import { deleteAllFileRecords } from './files'
-import { ensureFsStorageDirectory } from './fs'
-import { initLogger } from './logs'
-import { clearMnemonicHash } from './mnemonic'
-import { reconnectIndexer, resetSdk } from './sdk'
-import { getHasOnboarded, setHasOnboarded } from './settings'
-import { ensureTempFsStorageDirectory } from './tempFs'
-import { cancelAllUploads } from './uploads'
 
 export type InitStep = {
   id: string
@@ -39,13 +8,13 @@ export type InitStep = {
   startedAt: number
 }
 
-type AppInitState = {
+type AppState = {
   steps: Map<string, InitStep>
   isInitializing: boolean
   initializationError: string | null
 }
 
-const useAppInitStore = create<AppInitState>(() => {
+export const useAppStore = create<AppState>(() => {
   return {
     steps: new Map<string, InitStep>(),
     isInitializing: true,
@@ -53,122 +22,10 @@ const useAppInitStore = create<AppInitState>(() => {
   }
 })
 
-const { setState } = useAppInitStore
-
-export async function initApp(): Promise<void> {
-  startInitState()
-
-  const hasOnboarded = await getHasOnboarded()
-
-  const steps: StepDefinition[] = [
-    {
-      id: 'prepare',
-      label: 'Starting application',
-      message: 'Initializing...',
-      runner: async () => {
-        ensureFsStorageDirectory()
-        ensureTempFsStorageDirectory()
-        // Migrate keychain items to use AFTER_FIRST_UNLOCK accessibility
-        // so they can be accessed in background mode.
-        await migrateKeychainAccessibility()
-      },
-    },
-    {
-      id: 'migrations',
-      label: 'Initializing database',
-      message: 'Updating schema...',
-      runner: async (updateDetail) => {
-        await initializeDB({
-          onProgress: (event) => {
-            updateDetail(event.message)
-          },
-        })
-        await initLogger()
-      },
-    },
-  ]
-
-  if (hasOnboarded) {
-    steps.push({
-      id: 'connect',
-      label: 'Connecting to indexer',
-      message: 'Initializing SDK...',
-      runner: async () => {
-        const connected = await reconnectIndexer()
-        if (!connected) {
-          throw new Error('Failed to connect to indexer.')
-        }
-      },
-    })
-  }
-
-  steps.push({
-    id: 'services',
-    label: 'Starting background services',
-    message: 'Launching background services...',
-    runner: async () => {
-      initUploadScanner()
-      initSyncDownEvents()
-      initSyncNewPhotos()
-      initSyncPhotosArchive()
-      initBackgroundTasks()
-      initSyncUpMetadata()
-      initThumbnailScanner()
-      initFsOrphanScanner()
-      initFsEvictionScanner()
-      await initLogRotation()
-    },
-  })
-
-  await runSteps(steps)
-
-  endInitState()
-}
-
-function cancelAllTransfers() {
-  cancelAllUploads()
-  cancelAllDownloads()
-}
-
-export async function shutdownApp() {
-  cancelAllTransfers()
-}
-
-export async function resetData() {
-  await deleteAllFileRecords()
-  await resetDb()
-  await resetSyncDownCursor()
-  cancelAllTransfers()
-}
-
-export async function resetApp() {
-  startInitState()
-  shutdownAllServiceIntervals()
-
-  await runSteps([
-    {
-      id: 'reset',
-      label: 'Resetting application',
-      message: 'Clearing data...',
-      runner: async () => {
-        await resetData()
-        await clearAppKeys()
-        await clearMnemonicHash()
-        await setHasOnboarded(false)
-        await resetSdk()
-        await resetPhotosNewCursor()
-        await resetPhotosArchiveCursor()
-      },
-    },
-  ])
-
-  await initApp()
-}
-
-// selectors
+export const { setState: setAppState } = useAppStore
 
 export const [getInitSteps, useInitSteps] = createGetterAndSelector(
-  useAppInitStore,
+  useAppStore,
   (state) => {
     const steps = Array.from(state.steps.values())
     return steps.sort((a, b) => a.startedAt - b.startedAt)
@@ -176,7 +33,7 @@ export const [getInitSteps, useInitSteps] = createGetterAndSelector(
 )
 
 export const [getCurrentInitStep, useCurrentInitStep] = createGetterAndSelector(
-  useAppInitStore,
+  useAppStore,
   (state) => {
     const steps = Array.from(state.steps.values())
     return steps.sort((a, b) => a.startedAt - b.startedAt).at(-1)
@@ -184,91 +41,14 @@ export const [getCurrentInitStep, useCurrentInitStep] = createGetterAndSelector(
 )
 
 export const [getInitializationError, useInitializationError] =
-  createGetterAndSelector(useAppInitStore, (state) => state.initializationError)
+  createGetterAndSelector(useAppStore, (state) => state.initializationError)
 
 export const [getIsInitializing, useIsInitializing] = createGetterAndSelector(
-  useAppInitStore,
+  useAppStore,
   (s) => s.isInitializing,
 )
 
 export const [getShowSplash, useShowSplash] = createGetterAndSelector(
-  useAppInitStore,
+  useAppStore,
   (s) => s.isInitializing || s.initializationError,
 )
-
-// helpers
-
-function startInitState(): void {
-  setState({
-    steps: new Map<string, InitStep>(),
-    isInitializing: true,
-    initializationError: null,
-  })
-}
-
-function endInitState(): void {
-  setState({
-    steps: new Map<string, InitStep>(),
-    isInitializing: false,
-    initializationError: null,
-  })
-}
-
-function nextInitStep(step: Omit<InitStep, 'startedAt'>): void {
-  const nextStep = {
-    ...step,
-    startedAt: Date.now(),
-  }
-  setState((state) => {
-    const newSteps = new Map(state.steps)
-    newSteps.set(step.id, nextStep)
-    return { steps: newSteps }
-  })
-}
-
-function updateInitStep(step: { id: string; message?: string }): void {
-  setState((state) => {
-    const prev = state.steps.get(step.id)
-    if (!prev) return state
-    const next: InitStep = {
-      ...prev,
-      ...step,
-    }
-    const newSteps = new Map(state.steps)
-    newSteps.set(step.id, next)
-    return { steps: newSteps }
-  })
-}
-
-type StepDefinition = {
-  id: string
-  label: string
-  message: string
-  runner: (updateMessage: (message: string) => void) => Promise<void>
-}
-
-async function runSteps(steps: StepDefinition[]): Promise<void> {
-  for (const stepDef of steps) {
-    nextInitStep({
-      id: stepDef.id,
-      label: stepDef.label,
-      message: stepDef.message,
-    })
-
-    const updateMessage = (message: string) => {
-      updateInitStep({
-        id: stepDef.id,
-        message,
-      })
-    }
-
-    try {
-      await stepDef.runner(updateMessage)
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error)
-      updateMessage(message)
-      setState((state) => ({ ...state, initializationError: message }))
-      return // Stop execution on error
-    }
-  }
-}
