+++ src/stores/librarySwr.ts
@@ -1,0 +1,3 @@
+import { buildSWRHelpers } from '../lib/swr'
+
+export const librarySwr = buildSWRHelpers('library')
